refactor(nat-lab): group client code into a libtelio_client package (LLT-5791)
Collect the libtelio Client and its facades into a proper tests/libtelio_client/
package with clean module names:

  telio.py            -> libtelio_client/client.py
  runtime.py          -> libtelio_client/runtime.py
  client_log.py       -> libtelio_client/log.py
  client_events.py    -> libtelio_client/events.py
  client_vpn.py       -> libtelio_client/vpn.py
  client_analytics.py -> libtelio_client/analytics.py
  client_tp_lite.py   -> libtelio_client/tp_lite.py

__init__.py re-exports Client/Runtime/Events/WontHappenError, so external code
uses 'from tests.libtelio_client import Client' (12 call sites updated; no
backward-compat shim). Intra-package imports are package-qualified.

log_collector.py and utils/diagnostics.py stay outside the package: they're
shared test infrastructure used by non-client code (conftest, connection_util).
