+++ nat-lab/tests/libtelio_client/events.py
@@ -9,7 +9,7 @@ from tests.utils.logger import log
 from typing import List, Optional, TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from tests.telio import Client
+    from tests.libtelio_client.client import Client
 
 
 class ClientEvents:
