+++ nat-lab/tests/libtelio_client/log.py
@@ -3,7 +3,7 @@ from tests.log_collector import get_log_without_flush
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from tests.telio import Client
+    from tests.libtelio_client.client import Client
 
 
 class ClientLog:
