+++ nat-lab/tests/libtelio_client/vpn.py
@@ -4,7 +4,7 @@ from tests.utils.bindings import LinkState, NodeState, PathType
 from typing import Optional, TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from tests.telio import Client
+    from tests.libtelio_client.client import Client
 
 
 class ClientVpn:
