+++ nat-lab/tests/telio_test.py
@@ -2,7 +2,7 @@ import asyncio
 import copy
 import pytest
 import time
-from tests.runtime import Runtime, Events, WontHappenError
+from tests.libtelio_client import Runtime, Events, WontHappenError
 from tests.utils.asyncio_util import run_async_contexts, run_async_context
 from tests.utils.bindings import (
     NodeState,
