+++ nat-lab/tests/test_mesh_exit_through_peer.py
@@ -3,8 +3,8 @@ import pytest
 from contextlib import AsyncExitStack
 from tests import config
 from tests.helpers import setup_api, setup_mesh_nodes, SetupParameters
+from tests.libtelio_client import Client
 from tests.mesh_api import API
-from tests.telio import Client
 from tests.utils import testing, stun
 from tests.utils.bindings import (
     default_features,
